lookup SocketKind with sock.type
Confusingly `sock.proto` is not analogous to the PROXY protocol's socket
protocol, rather the `sock.type` attribute is. In almost all cases
`sock.proto` will be zero as per
https://docs.python.org/3/library/socket.html#socket.socket

This was resulting in the command `proxyprotocol-server --service
localhost:8082 localhost:8080` always injecting a PROXY protocol header
with a protocol of int 0 to the downstream server. This PR fixes that,
it now correctly injects int 1 in the above scenario and the other
`SocketKind` lookups should also work correctly.